% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actuarial.R
\docType{data}
\name{actuarial}
\alias{actuarial}
\title{Actuarial}
\description{
actuarial data
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{year} 
\item\code{payroll} 
\item\code{claims} 
}
}

