% Generated by roxygen2: do not edit by hand
\docType{data}
\name{afghan.deaths}
\alias{afghan.deaths}
\title{afghan.deaths}
\description{
NATO fatalities in Afghanistan.  see page 365
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{x3} 
\item\code{x5} 
\item\code{x4} 
}
<<<<<<< HEAD
=======

The variables included in the dataset are:
\itemize{
\item\code{N} 
\item\code{dur} 
}

Row names indicate country.
References: Browne, E. C., Frendreis, J. P., and Gleiber, D. W. (1986). The Process of Cabinet Dissolution: An Exponential Model of Duration and Stability in Western Democracies. American Journal of Political Science 30, 628-650.
>>>>>>> origin/master
}

