% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baldus-data.R
\docType{data}
\name{baldus}
\alias{baldus}
\title{baldus}
\source{
Baldus, D. C., Pulaski, C., & Woodworth, G. (1983). Comparative review of death sentences: An empirical study of the Georgia experience. The Journal of Criminal Law and Criminology (1973-), 74(3), 661-753.
}
\usage{
data(baldus)
}
\description{
Data from Baldus Study on death sentences in Georgia (Exercise 14.2, p. 521).
To use the data in JAGS or WinBugs, see \code{baldus.jags} and \code{balfus.winbugs}, respectively.
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{race} Defendant's race
\item\code{educatn} Defendant's education
\item\code{employm}
\item\code{SES}
\item\code{married}
\item\code{num.chld}
\item\code{military}
\item\code{pr.arrst}
\item\code{plea}
\item\code{sentence}
\item\code{defense}
\item\code{dp.sght}
\item\code{jdge.dec}
\item\code{pen.phse}
\item\code{did.appl}
\item\code{out.appl}
\item\code{vict.age} Victim's age
\item\code{vict.rel} 
\item\code{vict.st1}
\item\code{vict.st2}
\item\code{specialA}
\item\code{methodA}
\item\code{num.kill}
\item\code{num.prps}
\item\code{def.age} Defendant's age
\item\code{aggrevat} Aggravating circumstances
\item\code{bloody}
\item\code{fam.lov}
\item\code{insane} Defendant invoked insanity defense
\item\code{mitcir}
\item\code{num.depr}
\item\code{rape} Rape involved
}
}
\seealso{
baldus.jags baldus.winbugs
}

