% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expo.metrop.R
\name{biv.exp}
\alias{biv.exp}
\title{expo.metrop}
\usage{
expo.metrop(m,x,y,L1,L2,L,B)
}
\arguments{
\item{x}{starting point for the x vector}

\item{y}{starting point for the y vector}

\item{L1}{event intensity for the x dimension}

\item{L2}{event intensity for the y dimension}

\item{L}{shared event intensity}

\item{m}{number of iterations}

\item{B}{upper bound}
}
\description{
Simple Metropolis algorithm demonstration using a bivariate exponential target from Chapter 1 (pages 33-37).
}
\author{
Jeff Gill
}

