% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/child.R
\docType{data}
\name{child}
\alias{child}
\title{child}
\source{
Meier, K.J. and Keisler, L.R. (1996). Public Administration as a Science of the Artificial: A Method for Prescription, Public Administration Review 56, 459-466.
}
\usage{
child
}
\description{
Child Support Collection Policies. See page 128
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{SCCOLL} Change in Child Support collections
\item\code{ACES} Chapters per Population
\item\code{INSTABIL} Policy Instability
\item\code{AAMBIG} Policy Ambiguity
\item\code{CSTAFF} Change in Agency Staffing
\item\code{ARD} State Divorce Rate
\item\code{ASLACK} Organizational Slack
\item\code{AEXPEND} State Level Expenditures
}
}

