% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contracep.R
\docType{data}
\name{contracep}
\alias{contracep}
\title{contracep}
\source{
Wong, G. Y. and Mason, W. M. (1985).	The Hierarchical Logistic Regression Model for Multilevel Analysis.	Journal of the American Statistical Association 80, 513-524.
}
\usage{
data(contracep)
}
\description{
Contraception Data by country. See page 446
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{Country} Developing countries by size
\item\code{URC}
\item\code{WED}
\item\code{FPE}
\item\code{WED.FPE}
}
}

