% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\docType{data}
\name{dp}
\alias{dp}
\title{dp}
\source{
Norrander, B. (2000).	The Multi-Layered Impact of Public Opinion on Capital Punishment Implementation in the American States. Political Research Quarterly 53, 771-793.
}
\usage{
data(dp)
}
\description{
Death Penalty Data, See Page 189.
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{X} 
\item\code{EXECUTIONS} 
\item\code{INCOME} 
\item\code{PERPOVERTY} 
\item\code{PERBLACK} 
\item\code{VC100k96} 
\item\code{SOUTH} 
\item\code{PROPDEGREE} 
}
}

