% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicspend.R
\docType{data}
\name{elicspend}
\alias{elicspend}
\title{elicspend}
\usage{
data(elicspend)
}
\description{
Campaign spending data. See page 164
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{X400} 
\item\code{X2500} 
\item\code{X4000} 
}
}

