% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnic.immigration.R
\docType{data}
\name{ethnic.immigration}
\alias{ethnic.immigration}
\title{ethnic.immigration}
\source{
Peach, C. (1997). Postwar Migration to Europe: Reflux, Influx, Refuge.	Social Science Quarterly 78, 269-283.
}
\usage{
data(ethnic.immigration)
}
\description{
Ethnic Immigration data. see page 280.
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{Country.of.Origin} 
\item\code{Estimated.Total.K.} 
\item\code{Percent.of.Total} 
}
}

