% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.known.var.R
\name{n.post1}
\alias{n.post1}
\title{norm.known.var}
\usage{
n.post1(data.vec,pop.var,prior.mean,prior.var)
}
\arguments{
\item{data.vec}{a vector of assumed normally distributed data}

\item{pop.var}{known population variance}

\item{prior.mean}{mean of specified prior distribution for mu}

\item{prior.var}{variance of specified prior distribution for mu}
}
\description{
A function to calculate posterior quanties for a normal-normal model with known variance (pages 74-77).
It produces the posterior mean, variance, and 95\% credible interval for user-specified prior.
}
\author{
Jeff Gill
}

