<<<<<<< Updated upstream
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> Stashed changes
% Please edit documentation in R/nc.sub.dat - data.R
\docType{data}
\name{nc.sub.dat}
\alias{nc.sub.dat}
\title{nc.sub.dat}
\usage{
nc.sub.dat
}
\description{
North Carolina county level health data from the 2000 U.S. census and North Carolina public records, see page 78.

The variables included in the dataset are:
\itemize{
\item\code{Substantiated.Abuse} within family documented abuse for the county
\item\code{Percent.Poverty} percent within the county living in poverty, U.S. definition
\item\code{Total.Population} county population/1000

}
}

