% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opic-data.R
\docType{data}
\name{opic}
\alias{opic}
\title{opic}
\usage{
data(opic)
}
\description{
private capital investment data. See Page 390.

The variables included in the dataset are:
\itemize{
\item\code{Fund} Name of the private company
\item\code{Age} Years the company has been in existence
\item\code{Status} Whether the company is investing or divesting
\item\code{Size} Size of the investment made
}
}

