% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.walk.MH.R
\name{plot.walk.MH}
\alias{plot.walk.MH}
\title{plot.walk.MH}
\usage{
plot.walk.MH(walk.mat)
}
\arguments{
\item{walk.mat}{}
}
\description{
Plot.walk.MH code used to produce figure 10.4
}
\author{
Jeff Gill
}

