% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recidivism-data.R
\docType{data}
\name{Recidivism statistics}
\alias{Recidivism statistics}
\title{recidivism}
\usage{
data(recidivism)
}
\description{
Recidivism Rates. See page 207

The variables included in the dataset are:
\itemize{
\item\code{Crime.Type} The type of crime committed
\item\code{Released} The number of individuals released from a facility
\item\code{Returned} The number of individuals returned to a facility
\item\code{Percentage} (The number of individuals returned to a facility)/(The number of individuals released from a facility)
}
}

