% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retail.sales-data.R
\docType{data}
\name{Retail Sales}
\alias{Retail Sales}
\title{retail.sales}
\usage{
data(retail.sales)
}
\description{
The variables included in the dataset are:
\itemize{
\item\code{Time} 
\item\code{DSB} 
\item\code{EMP} 
\item\code{BDG} 
\item\code{CAR}
\item\code{FRN}
\item\code{GMR}
}
}

