% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmultnorm.R
\name{rmultnorm}
\alias{rmultnorm}
\title{rmultnorm}
\usage{
rmultnorm(n, mu, vmat, tol = 1e-07)
}
\arguments{
\item{n}{XXX}

\item{mu}{XXX}

\item{vmat}{XXX}

\item{tol}{XXX}
}
\description{
a function to generate random multivariate Gaussians.
}
\author{
Jeff Gill
}
\seealso{
\code{\link{biv.norm.post}}
}

