% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/durations.hpd.R
\name{simple.hpd}
\alias{simple.hpd}
\title{durations.hpd}
\usage{
simple.hpd(support,fn.eval,start,stop,target=0.90,tol=0.01)
}
\arguments{
\item{support}{x-axis values}

\item{function}{values at x-axis points}

\item{starting}{point in the vectors}

\item{stoppng}{point in the vectors}

\item{Desired}{X Level}

\item{Tolerance}{for round-off}
}
\description{
Simple HPD calculator from Chapter 2 (page 51).
}
\examples{
## Not run:
{
  data(cabinet.duration)
  ruler <- seq(0.45,0.75,length=10000)
  g.vals <- round(dgamma(ruler,shape=sum(N), rate=sum(N*dur)),2)
  start.point  <- 1000; stop.point <- length(g.vals)
  simple.hpd(ruler,g.vals,start.point,stop.point)
}

## End(Not run)
}
\author{
Jeff Gill
}

