% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strikes-data.R
\docType{data}
\name{Strikes}
\alias{Strikes}
\title{strikes}
\usage{
data(strikes)
}
\description{
THIS NEEDS TO BE FILLED IN

The variables included in the dataset are:
\itemize{
\item\code{Year} The year the strikes occurred
\item\code{Counts} The number of strikes that occurred per year
}
}

