% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrorism-data.R
\docType{data}
\name{Terrorism Comparisons Using Statistics}
\alias{Terrorism Comparisons Using Statistics}
\title{terrorism}
\usage{
data(terrorism)
}
\description{
Dataset comparing incidents of terrorism to car accidents, suicide, and murder

The variables included in the dataset are:
\itemize{
\item\code{Year} The given year in which the statistics occurred
\item\code{X.Terrorism} The number of terrorist attacks that would occur per 100000 in the given year
\item\code{X.Car.Accidents} The number of car accidents that would occur per 100000 in the given year
\item\code{X.Suicide} The number of suicide that would occur per 100000 in the given year
}
}

