% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texas-data.R
\docType{data}
\name{texas}
\alias{texas}
\title{texas}
\usage{
data(texas)
}
\description{
Poverty in Texas

The variables included in the dataset are:
\itemize{
\item\code{} 
\item\code{} 
\item\code{} 
\item\code{} 
}
}

