% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk.G.R
\name{walk}
\alias{walk}
\title{walk.G}
\usage{
## S3 method for class 'G'  
   plot.walk.G(walk.mat,sim.rm,X=1,Y=2)
}
\arguments{
\item{walk.mat}{walk.mat}

\item{sim.rm}{sim.rm}

\item{X}{X}

\item{Y}{Y}
}
\description{
Plot.walk.g code used to produce figure 10.4
}
\author{
Jeff Gill
}

