% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china.wars.R
\docType{data}
\name{wars}
\alias{wars}
\title{china.wars}
\format{An object of class \code{data.frame} with 104 rows and 14 columns.}
\usage{
wars
}
\description{
Modeling code for the example of ancient Chinese wars. See page 125-127
}
\author{
Jeff Gill
}
\keyword{datasets}

