% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actuarial.R
\docType{data}
\name{actuarial}
\alias{actuarial}
\title{actuarial}
\source{
Scollnik, D. P. M. (2001). Actuarial Modeling with MCMC and BUGS. North American Actuarial Journal 5, 95-124.
}
\usage{
data(actuarial)
}
\description{
Actuarial data
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{year} 
\item\code{payroll} 
\item\code{claims} 
}
}

