% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afghan.deaths.R
\docType{data}
\name{afghan.deaths}
\alias{afghan.deaths}
\title{afghan.deaths}
\description{
NATO fatalities in Afghanistan.  see page 365
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{x3} 
\item\code{x5} 
\item\code{x4} 
}
}

