% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa.R
\docType{data}
\name{africa}
\alias{africa}
\title{africa}
\description{
African Coups Data, pp.551-552
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{MILTCOUP} 
\item\code{MILITARY} 
\item\code{POLLIB}
\item\code{PARTY93}
\item\code{PCTVOTE}
\item\code{PCTTURN}
\item\code{SIZE}
\item\code{POP}
\item\code{NUMREGIM}
\item\code{NUMELEC}
}
}

