% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strikes-data.R
\docType{data}
\name{French Labor Strikes}
\alias{French Labor Strikes}
\title{strikes}
\format{data frame with 11 observations of strikes that occurred in different years with 1 explanatory variable}
\source{
Conell, C. and Cohn, S. (1995). Learning from Other People’s Actions: Environmental Variation and Diffusion in French Coal Mining Strikes, 1890-1935. American Journal of Sociology 101, 366-403.
}
\usage{
data(strikes)
}
\description{
French Coal Strikes, see page 212 and 213

The variables included in the dataset are:
\itemize{
\item\code{Year} The year the labor strikes in France occurred
\item\code{Counts} The number of labor strikes that occurred in France per year
}
}

