% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strikes-data.R
\docType{data}
\name{French Labor Strikes}
\alias{French Labor Strikes}
\title{strikes}
\usage{
data(strikes)
}
\description{
French Coal Strikes, see page 212 and 213

The variables included in the dataset are:
\itemize{
\item\code{Year} The year the labor strikes in France occurred
\item\code{Counts} The number of labor strikes that occurred in France per year
}
}

