% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/military-data.R
\docType{data}
\name{military}
\alias{military}
\title{military}
\usage{
data(military)
}
\description{
A dataset of two variables. The proportional changes in military personnel for the named countries. See pages 79-86

The variables included in the dataset are:
\itemize{
\item\code{Year} The year selected to evaluate
\item\code{Yugoslavia} The proportion change in the size of Yugoslavia's military 
\item\code{Albania} The proportion change in the size of Albania's military 
\item\code{Bulgaria} The proportion change in the size of Bulgaria's military 
\item\code{Czechoslovakia} The proportion change in the size of Czechoslovakia's military 
\item\code{German.Dem.Republic} The proportion change in the size of the German Democratic Republic's military
\item\code{Hungary} The proportion change in the size of Hungary's military 
\item\code{Poland} The proportion change in the size of Poland's military 
\item\code{Rumania} The proportion change in the size of Romania's military 
\item\code{USSR} The proportion change in the size of the Soviet Union's military 
 
}
}

