% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.sub.dat - data.R
\docType{data}
\name{nc.sub.dat}
\alias{nc.sub.dat}
\title{nc.sub.dat}
\usage{
nc.sub.dat
}
\description{
North Carolina county level health data from the 2000 U.S. census and North Carolina public records.

The variables included in the dataset are:
\itemize{
\item\code{Substantiated.Abuse} Documented substantiated abuse cases within a county
\item\code{Percent.Poverty} Percent within the county living in poverty as per the U.S. definition
\item\code{Total.Population} (The total population observed)/1000
 
}
}

