% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/texas-data.R
\docType{data}
\name{texas}
\alias{texas}
\title{texas}
\usage{
data(texas)
}
\description{
Poverty in Texas, see page 299

The variables included in the dataset are:
\itemize{
\item\code{POV}
\item\code{BLK} the proportion of Black residents in the county
\item\code{LAT} the proportion of Latino residents in the county
\item\code{GVT} a dichotomous variable indicating whether government activities contributed a weighted annual average of 25% or more labor and proprietor income over the previous 3 years
\item\code{SVC} a dichotomous variable indicating whether service activities contributed a weighted annual average of 50% or more labor and proprietor income over the previous 3 years
\item\code{FED} a dichotomous variable indicating whether federally owned lands make up 30% or more of a county's land area
\item\code{XFR}
\item\code{POP} the log of the county population total for 1989
}
}

