% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texas-data.R
\docType{data}
\name{Texas}
\alias{Texas}
\title{texas}
\usage{
data(texas)
}
\description{
THERE NEEDS TO BE A DESCRIPTION HERE

The variables included in the dataset are:
\itemize{
\item\code{} 
\item\code{} 
\item\code{} 
\item\code{} 
}
}

