% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wars-data.R
\docType{data}
\name{wars}
\alias{wars}
\title{wars}
\usage{
data(wars)
}
\description{
Data for Chinese wars example

The variables included in the dataset are:
\itemize{
\item\code{CASEhex5FID} The case name assigned to each war 
\item\code{WARhex5FNAME} The name of the war  
\item\code{V1} Starting date of the war 
\item\code{V2} Ending date of the war 
}
}

