% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wars-data.R
\docType{data}
\name{wars}
\alias{wars}
\title{wars}
\format{a data frame of 104 observations of different China wars with 15 explanatory variables}
\source{
Cioffi-Revilla, C. and Lai, D. (1995). War and Politics in Ancient China, 2700 B.C. to 722 B.C.: Measurement and Comparative Analysis. Journal of Conflict Resolution 39, 467-494.
}
\usage{
data(wars)
}
\description{
Data for Chinese wars example, see page 163

The variables included in the dataset are:
\itemize{
<<<<<<< Updated upstream
\item\code{ONSET} 
\item\code{TERM} 
\item\code{EXTENT} number of belligerents
\item\code{ETHNIC} intra-group or inter-group
\item\code{DIVERSE} number of ethnic groups participating as belligerents 
=======
\item\code{ONSET}
\item\code{TERM}
\item\code{EXTENT} number of belligerents
\item\code{ETHNIC} intra-group or inter-group
\item\code{DIVERSE} number of ethnic groups participating as belligerents
>>>>>>> Stashed changes
\item\code{ALLIANCE} total number of alliances among belligerents
\item\code{DYADS} number of alliance pairs
\item\code{POL.LEV}
\item\code{COMPLEX}
\item\code{POLAR} number of relatively major or great powers at the time of onset
\item\code{BALANCE} the difference in military capabilities: minor-minor, minor-major, major-major
\item\code{TEMPOR} type of war: protracted rivalry, integrative conquest, disintegrative/fracturing conflict, sporadic event
<<<<<<< Updated upstream
\item\code{SCOPE} 
=======
\item\code{SCOPE}
>>>>>>> Stashed changes
\item\code{DURATION} duration of conflict, measured in years
}
}

