% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china.wars.R
\docType{data}
\name{wars}
\alias{wars}
\title{china.wars}
\format{An object of class \code{data.frame} with 104 rows and 14 columns.}
\source{
Claudio Cioffi-Revilla and David Lai, 2001, \cr
 "Chinese Warfare and Politics in the Ancient East Asian International System",\cr
  hdl:1902.1/02016 \url{http://id.thedata.org/hdl\%3A1902.1\%2F02016 Social Science Program} \cr
   Download from \url{http://dvn.iq.harvard.edu/dvn/dv/mra/faces/study/StudyPage.jsp?studyId=110&studyListingIndex=0_9aa809933d3c24eb76216fa477f8} \cr
    Henry A. Murray Research Archive \cr
     Center for International Relations, Department of Political Science, University of Colorado, Boulder, USA
}
\usage{
wars
}
\description{
Modeling code for the example of ancient Chinese wars. See page 125-127
}
\author{
Jeff Gill
}
\keyword{datasets}

